requirements.txt
openai
toolz
Pillow
requests
python-pptx
python-dotenv
